fix: circual deps
