+++ lib/shared/config-manager/src/index.ts
@@ -1,4 +1,4 @@
-import { DVCLogger } from '@devcycle/types'
+import { DVCLogger, DevCycleOptions } from '@devcycle/types'
 import { getEnvironmentConfig } from './request'
 import { ResponseError, UserError } from '@devcycle/server-request'
