+++ lib/shared/types/src/index.ts
@@ -1,6 +1,7 @@
 export * from './logger'
 export * from './reporter'
 export * from './types/apis/sdk/clientSDKAPI'
+export * from './types/apis/sdk/serverSDKTypes'
 export * from './types/apis/sdk/track'
 export * from './types/config/configBody'
 export * from './types/validators/dvcCustomDataJSON'
