+++ sdk/js-cloud-server/src/request.ts
@@ -1,6 +1,7 @@
-import { post } from '@devcycle/server-request'
 import { DVCPopulatedUser } from './models/populatedUser'
-import { DevCycleEvent, DevCycleOptions } from './types'
+import { DevCycleEvent } from './types'
+import { DevCycleOptions } from '@devcycle/types'
+import { post } from '@devcycle/server-request'
 
 export const HOST = '.devcycle.com'
